Cleaned up code and modernized C++ usage

+++ source/gui_main.cpp
@@ -40,32 +40,34 @@ void GuiMain::draw() {
   float x = 0, y = 10, currItem = 0;
   float selectedX = 0, selectedY = 0;
 
-  for(auto title : Title::g_titles) {
+  for (auto title : Title::g_titles) {
     Gui::drawImage(x - xOffset, y, 256, 256, title.second->getTitleIcon(), IMAGE_MODE_RGB24);
     Gui::drawShadow(x - xOffset, y, 256, 256);
 
-    if(currItem == m_selected.titleIndex) {
+    if (currItem == m_selected.titleIndex) {
       selectedX = x - xOffset;
       selectedY = y;
       m_selected.titleId = title.first;
     }
 
     y = y == 10 ? 266 : 10;
-    if(y == 10)
+
+    if (y == 10)
       x += 256;
 
     currItem++;
   }
 
-  if(selectionState >= TITLE_SELECT) {
+  if (selectionState >= TITLE_SELECT) {
     Gui::drawRectangled(selectedX - 10, selectedY - 10, 276, 276, selectionState == TITLE_SELECT ? currTheme.highlightColor : currTheme.selectedColor);
     Gui::drawRectangled(selectedX - 5, selectedY - 5, 266, 266, currTheme.selectedButtonColor);
     Gui::drawImage(selectedX, selectedY, 256, 256, Title::g_titles[m_selected.titleId]->getTitleIcon(), IMAGE_MODE_RGB24);
     Gui::drawShadow(selectedX - 10, selectedY - 10, 276, 276);
   }
 
-  if(selectionState >= ACCOUNT_SELECT && Title::g_titles[m_selected.titleId]->getUserIDs().size() > 0) {
-      for(u8 i = 0; i < Title::g_titles[m_selected.titleId]->getUserIDs().size(); i++)
+  if (selectionState >= ACCOUNT_SELECT && Title::g_titles[m_selected.titleId]->getUserIDs().size() > 0) {
+
+      for (u8 i = 0; i < Title::g_titles[m_selected.titleId]->getUserIDs().size(); i++)
         Gui::drawShadow(40 + i * 150, 560, 128, 128);
 
       Gui::drawRectangled(40 + m_selected.accountIndex * 150 - 10, 550, 148, 148, currTheme.highlightColor);
@@ -73,14 +75,15 @@ void GuiMain::draw() {
       Gui::drawShadow(40 + m_selected.accountIndex * 150 - 10, 550, 148, 148);
 
       u16 accountX = 0;
-      for(u128 userID : Title::g_titles[m_selected.titleId]->getUserIDs()) {
+
+      for (u128 userID : Title::g_titles[m_selected.titleId]->getUserIDs()) {
         Gui::drawImage(40 + accountX, 560, 128, 128, Account::g_accounts[userID]->getProfileImage(), IMAGE_MODE_RGB24);
         accountX += 150;
       }
   }
 
-  if(xOffset != xOffsetNext) {
-    if(xOffsetNext > xOffset)
+  if (xOffset != xOffsetNext) {
+    if (xOffsetNext > xOffset)
       xOffset += ceil((abs(deltaOffset) > scrollSpeed) ? scrollSpeed : deltaOffset);
     else
       xOffset += floor((abs(deltaOffset) > scrollSpeed) ? scrollSpeed : deltaOffset);
@@ -91,48 +94,48 @@ void GuiMain::draw() {
 }
 
 void GuiMain::onInput(u32 kdown) {
-  if(kdown & KEY_LEFT) {
-    if(selectionState == TITLE_SELECT) {
-      if(((s16)m_selected.titleIndex - 2) >= 0)
+  if (kdown & KEY_LEFT) {
+    if (selectionState == TITLE_SELECT) {
+      if (static_cast<s16>(m_selected.titleIndex - 2) >= 0)
         m_selected.titleIndex -= 2;
-    } else if(selectionState == ACCOUNT_SELECT) {
-      if(((s16)m_selected.accountIndex - 1) >= 0)
+    } else if (selectionState == ACCOUNT_SELECT) {
+      if (static_cast<s16>(m_selected.accountIndex - 1) >= 0)
         m_selected.accountIndex--;
     }
   }
 
-  if(kdown & KEY_RIGHT) {
-    if(selectionState == TITLE_SELECT) {
-      if((u16)(m_selected.titleIndex + 2) < Title::g_titles.size())
+  if (kdown & KEY_RIGHT) {
+    if (selectionState == TITLE_SELECT) {
+      if (static_cast<u16>(m_selected.titleIndex + 2) < Title::g_titles.size())
         m_selected.titleIndex += 2;
-    } else if(selectionState == ACCOUNT_SELECT) {
-      if((u16)(m_selected.accountIndex + 1) < Title::g_titles[m_selected.titleId]->getUserIDs().size())
+    } else if (selectionState == ACCOUNT_SELECT) {
+      if (static_cast<u16>(m_selected.accountIndex + 1) < Title::g_titles[m_selected.titleId]->getUserIDs().size())
         m_selected.accountIndex++;
     }
   }
 
-  if(kdown & KEY_UP || kdown & KEY_DOWN) {
-    if(selectionState == TITLE_SELECT) {
-      if((m_selected.titleIndex % 2) == 0) {
-        if((u16)(m_selected.titleIndex + 1) < Title::g_titles.size())
+  if (kdown & KEY_UP || kdown & KEY_DOWN) {
+    if (selectionState == TITLE_SELECT) {
+      if ((m_selected.titleIndex % 2) == 0) {
+        if (static_cast<u16>(m_selected.titleIndex + 1) < Title::g_titles.size())
           m_selected.titleIndex++;
       }
       else m_selected.titleIndex--;
     }
   }
 
-  if(kdown & KEY_A) {
-    if(selectionState == TITLE_SELECT)
+  if (kdown & KEY_A) {
+    if (selectionState == TITLE_SELECT)
       selectionState = ACCOUNT_SELECT;
-    else if(selectionState == ACCOUNT_SELECT && Title::g_titles[m_selected.titleId]->getUserIDs().size() > 0) {
+    else if (selectionState == ACCOUNT_SELECT && Title::g_titles[m_selected.titleId]->getUserIDs().size() > 0) {
       Title::g_currTitle = Title::g_titles[m_selected.titleId];
       Account::g_currAccount = Account::g_accounts[Title::g_titles[m_selected.titleId]->getUserIDs()[m_selected.accountIndex]];
       Gui::g_nextGui = GUI_EDITOR;
     }
   }
 
-  if(kdown & KEY_B) {
-    if(selectionState == ACCOUNT_SELECT) {
+  if (kdown & KEY_B) {
+    if (selectionState == ACCOUNT_SELECT) {
         selectionState = TITLE_SELECT;
         m_selected.accountIndex = 0;
     }
@@ -140,14 +143,14 @@ void GuiMain::onInput(u32 kdown) {
 }
 
 void GuiMain::onTouch(touchPosition &touch) {
-  switch(selectionState) {
+  switch (selectionState) {
     case TITLE_SELECT: {
-      u8 x = floor((float)(touch.px + xOffset) / 256.0F);
-      u8 y = floor((float)touch.py / 256.0F);
+      u8 x = floor((touch.px + xOffset) / 256.0F);
+      u8 y = floor(touch.py / 256.0F);
       u8 title = y + x * 2;
 
-      if(y <= 1 && title < Title::g_titles.size()) {
-        if(m_selected.titleIndex == title) {
+      if (y <= 1 && title < Title::g_titles.size()) {
+        if (m_selected.titleIndex == title) {
           Title::g_currTitle = Title::g_titles[m_selected.titleId];
           selectionState = ACCOUNT_SELECT;
         }
@@ -158,8 +161,8 @@ void GuiMain::onTouch(touchPosition &touch) {
     case ACCOUNT_SELECT: {
       u8 account = floor((touch.px - 40) / 150.0F);
 
-      if(account < Title::g_titles[m_selected.titleId]->getUserIDs().size() && touch.py > 560 && touch.py < (560 + 128)) {
-        if(m_selected.accountIndex == account) {
+      if (account < Title::g_titles[m_selected.titleId]->getUserIDs().size() && touch.py > 560 && touch.py < (560 + 128)) {
+        if (m_selected.accountIndex == account) {
           m_selected.userId = Title::g_titles[m_selected.titleId]->getUserIDs()[account];
           Title::g_currTitle = Title::g_titles[m_selected.titleId];
           Account::g_currAccount = Account::g_accounts[Title::g_titles[m_selected.titleId]->getUserIDs()[m_selected.accountIndex]];
